+++ CHANGELOG.md
@@ -6,6 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [0.2.0] - [UNRELEASED]
 ### Added
+- EligibilityData
+- EligibilityBenefit
+- EligibilityBenefits
+
+### Changed
+- Eligibility.query returns EligibilityData object
 
 ## [0.1.0] - 2020-03-04
 ### Added